+++ main.cpp
@@ -527,7 +527,7 @@ int main(){
                     DTEstadisticaEstudiante estadisticas = contUsuario.estadisticasEstudiante(estu);
 
                     imprimir("Estadisticas de "+ estu + ":" );
-                    map<string, float> porcentajesCursos = estadisticas.getPorcentajesCursos();
+                    map<string, int> porcentajesCursos = estadisticas.getPorcentajesCursos();
                     for (const auto& estads : porcentajesCursos) {
                         imprimir("Curso: " + estads.first + ", Porcentaje: " + to_string(estads.second) + "%");
                     }
